+++ src/command/scroll.rs
@@ -6,11 +6,17 @@ pub enum ScrollCommand {
 }
 
 impl ScrollCommand {
-    pub fn to_lines(self, page_height: i32) -> i32 {
+    pub fn to_lines(self, page_height: usize) -> i32 {
         match self {
             Self::Lines(n) => n,
-            Self::Pages(n) => n * page_height,
+            Self::Pages(n) => n * page_height as i32,
         }
+    }
+    /// compute the new scroll value
+    pub fn apply(self, scroll: usize, content_height: usize, page_height: usize) -> usize {
+        (scroll as i32 + self.to_lines(page_height))
+            .min(content_height as i32 - page_height as i32 + 1)
+            .max(0) as usize
     }
 }